requirements.txt
astroid==2.3.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
-e git+git@github.com:dwardu89/photo-sorter.git@c1eff97ba7582ae8b631f3f71a0437d42862cc3d#egg=photo_sorter
Pillow==7.0.0
pylint==2.4.4
six==1.13.0
typed-ast==1.4.0
wrapt==1.11.2
